requirements.txt
pandas
numpy
scikit-learn
requests
python-whois 
matplotlib
seaborn
jupyterlab
tldextract